cognito mapping hot fix
